+++ test/svelte3/integration/data/data.export.aliace.svelte
@@ -1,0 +1,11 @@
+<script>
+    /** 
+     * Description for variable that must be exported later 
+     * @type {Array<string>}
+     */
+    const classes = [ 'btn' ];
+
+    export { 
+        classes as class
+    };
+</script>