+++ test/svelte3/integration/data/data.export.many.svelte
@@ -1,0 +1,7 @@
+<script>
+    /* The `a` variable description */
+    let a = 1;
+    let b = 'Test';
+
+    export { a, b };
+</script>